+++ install/test/src/test_resource_detectors.cc
@@ -4,8 +4,10 @@
 #include <gtest/gtest.h>
 
 #include <opentelemetry/resource_detectors/container_detector.h>
+#include <opentelemetry/resource_detectors/container_detector_builder.h>
 #include <opentelemetry/resource_detectors/env_entity_detector.h>
 #include <opentelemetry/resource_detectors/process_detector.h>
+#include <opentelemetry/resource_detectors/process_detector_builder.h>
 #include <memory>
 
 TEST(ResourceDetectorsInstall, ContainerResourceDetector)
@@ -30,4 +32,24 @@ TEST(ResourceDetectorsInstall, ProcessResourceDetector)
       std::make_unique<opentelemetry::resource_detector::ProcessResourceDetector>();
   ASSERT_TRUE(detector != nullptr);
   ASSERT_NO_THROW(auto resource = detector->Detect());
+}
+
+TEST(ResourceDetectorsInstall, ContainerDetectorBuilder)
+{
+  auto builder = std::make_unique<opentelemetry::resource_detector::ContainerDetectorBuilder>();
+  ASSERT_TRUE(builder != nullptr);
+
+  opentelemetry::sdk::configuration::ContainerResourceDetectorConfiguration model;
+  auto detector = builder->Build(&model);
+  ASSERT_TRUE(detector != nullptr);
+}
+
+TEST(ResourceDetectorsInstall, ProcessDetectorBuilder)
+{
+  auto builder = std::make_unique<opentelemetry::resource_detector::ProcessDetectorBuilder>();
+  ASSERT_TRUE(builder != nullptr);
+
+  opentelemetry::sdk::configuration::ProcessResourceDetectorConfiguration model;
+  auto detector = builder->Build(&model);
+  ASSERT_TRUE(detector != nullptr);
 }